项目后端：
# 新增：

- 添加GoodSellingException
表示商品正在交易,不能改变商品信息

# 修改：

- 补充注释到GoodTagMapper
开发时不写注释就跟泡脚一样舒爽;回头维护自己亲手写出的屎山就像喝洗脚水一样难受

- 修改GoodsWithoutAccountAndSoldBO
现在GoodsWithoutAccountAndSoldBO不直接来源于数据库;而是从数据库中查出GoodBO,然后通过构造函数获得GoodsWithoutAccountAndSoldBO

- 完善卖家商品信息修改流程
现在当商品正在交易时,卖家不能进行商品信息修改或删除商品.

# 重要待办：

- 注释补齐！！！
- dto中对string类型的字段先验null在执行isBlank（）
- 新订单插入有问题
   - 问题: 商品因交易异常或交易取消后,不能对同一个商品再一次建立订单,因为goodId是唯一索引.
   - 目前解决方法: 取消goodId的唯一索引; 插入新bill时检查有无state为在交易的goodId.

+++ backend/gs_merge/src/main/java/com/gseek/gs/service/impl/SellerServiceImpl.java
@@ -7,6 +7,7 @@ import com.gseek.gs.dao.*;
 import com.gseek.gs.exce.ServerException;
 import com.gseek.gs.exce.business.ForbiddenException;
 import com.gseek.gs.exce.business.ParameterWrongException;
+import com.gseek.gs.exce.business.seller.GoodSellingException;
 import com.gseek.gs.pojo.bean.GoodPhotoFileBean;
 import com.gseek.gs.pojo.bean.GoodPhotoPathBean;
 import com.gseek.gs.pojo.bean.ParameterWrongBean;
@@ -53,6 +54,8 @@ public class SellerServiceImpl implements SellerService {
     OfferPriceMapper offerPriceMapper;
     @Autowired
     GoodCheckMapper goodCheckMapper;
+    @Autowired
+    BillMapper billMapper;
 
     @Override
     @Transactional(rollbackFor = Exception.class)
@@ -87,7 +90,11 @@ public class SellerServiceImpl implements SellerService {
         if (ownUserId != userId){
             throw new ForbiddenException();
         }
-        // todo 商品正在售出时禁止修改
+        // 商品正在售出时禁止修改
+        if (billMapper.selectBillByGoodId(dto.getGoodId()) != null ){
+            throw new GoodSellingException();
+        }
+
         GoodDO goodDO=new GoodDO(userName,dto);
         // 更新type和tag情况
         updateTypeAndTags(dto,goodDO);
@@ -105,7 +112,10 @@ public class SellerServiceImpl implements SellerService {
         if (userId != ownUserId){
             throw new ForbiddenException();
         }
-        // todo 商品正在售出时禁止修改
+        // 商品正在售出时禁止删除
+        if (billMapper.selectBillByGoodId(goodId) != null ){
+            throw new GoodSellingException();
+        }
         goodMapper.deleteGood(goodId);
 
         return result.gainDeleteSuccess();
@@ -130,14 +140,19 @@ public class SellerServiceImpl implements SellerService {
     }
 
     //todo 补充注释
+    // todo 下面这两个方法不要进行方法重载,因为这两个方法处理的业务不同
     /**
-     *
-     *
+     * 修改商品时,处理商品类型、tag、图片储存路径.
+     * 首先会清除商品对应的tag、图片路径信息.
+     * 商品类型: 数据库中有这种类型,则记录入good表中; 没有则抛出异常.
+     * tag: 数据库中没有这种tag会在tag表中新建,然后获取所有tag的tagId.最后一并记录入good_tag表中.
+     * 图片: 在minio中插入后,记录图片路径.
      * */
     private void updateTypeAndTags(PatchGoodsDTO dto,GoodDO goodDO) throws JsonProcessingException {
         int goodId=goodDO.getGoodId();
         String patchType=dto.getType();
-        if (patchType !=null){
+        //todo 为什么我要在这再验一遍参数???
+        if (patchType != null && !patchType.isBlank()){
             TagDO localType=tagMapper.selectTagByTagName(patchType);
             if (localType == null ){
                 throw new ParameterWrongException(new ParameterWrongBean()
@@ -175,23 +190,31 @@ public class SellerServiceImpl implements SellerService {
 
         goodCoverPicMapper.insertCoverPic(goodId,coverAndDetailPath.getCoverPaths());
         goodDetailPicMapper.insertDetailPic(goodId,coverAndDetailPath.getDetailPaths());
-
+        // 更新商品信息
+        goodMapper.updateGoodSelect(goodDO);
     }
     /**
+     * 新建商品时,处理商品类型、tag、图片储存路径.
+     * 商品类型: 数据库中有这种类型,则记录入good表中; 没有则抛出异常.
+     * tag: 数据库中没有这种tag会在tag表中新建,然后获取所有tag的tagId.最后一并记录入good_tag表中.
+     * 图片: 在minio中插入后,记录图片路径.
      *
-     *
+     * @param dto
+     * @param goodDO
+     * @param goodId 商品id
      * */
     private void updateTypeAndTags(PostGoodsDTO dto, GoodDO goodDO, int goodId) throws JsonProcessingException {
         String patchType=dto.getType();
-        if (patchType !=null){
-            TagDO localType=tagMapper.selectTagByTagName(patchType);
-            if (localType == null ){
-                throw new ParameterWrongException(new ParameterWrongBean()
-                        .addParameters("该商品类型不存在",patchType));
-            }
-            goodDO.setTypeTagId(localType.getTagId());
-            goodDO.setTypeTagName(localType.getTagText());
+
+        // 处理商品类型
+        TagDO localType=tagMapper.selectTagByTagName(patchType);
+        if (localType == null ){
+            throw new ParameterWrongException(new ParameterWrongBean()
+                    .addParameters("该商品类型不存在",patchType));
         }
+        goodDO.setTypeTagId(localType.getTagId());
+        goodDO.setTypeTagName(localType.getTagText());
+
 
         //修改tag
         List<TagDO> tags=new ArrayList<>();
@@ -199,13 +222,13 @@ public class SellerServiceImpl implements SellerService {
             for (String tagName:dto.getTag()){
                 tags.add(new TagDO(tagName));
             }
-            //数据库中没有的tag新建
-            if (tags.size() !=0){
+            // 如果数据库中没有该tag,会新建;已有则忽略插入操作
+            if (tags.size() !=0 ){
                 tagMapper.insertTags(tags);
             }
             //数据库中已有的tag查询
             for (TagDO tagDO:tags){
-                if (tagDO.getTagId()==null){
+                if (tagDO.getTagId() == null){
                     //todo 这太逆天了，就不能放在一次查吗
                     tagDO.setTagId(tagMapper.selectTagByTagName(tagDO.getTagText()).getTagId());
                 }
@@ -221,7 +244,9 @@ public class SellerServiceImpl implements SellerService {
 
         goodCoverPicMapper.insertCoverPic(goodId,coverAndDetailPath.getCoverPaths());
         goodDetailPicMapper.insertDetailPic(goodId,coverAndDetailPath.getDetailPaths());
-
+        // 更新商品信息
+        //todo 应该在插入新商品信息(71行)前向goodDO中加入type信息,而不是在这里.
+        goodMapper.updateGoodSelect(goodDO);
     }
